Refactor PreciseLanguageMeta
Made it more general so it can be used for any type of multiple choice question with fixed questions and set of answers. Following that the file was renamed and moved.

The generators using that question type were refactored to reflect the changes and renamed to fit the naming scheme.