Rewrite graph rendering.
The functionality was extended to support directed
and weighted graphs.
Also removed d3 and used React components instead.
For this to work, the graph needs to know where to place the nodes.
So a generator for grid graphs was added.
The string format of a graph was changes to support the additional information.