Make huffman codes more flexible.
They allow now for arbitrary labels on the edges of a huffman tree. In turn, we don't
need special rules for the case that some nodes have the same weight.
The only restriction now is, that the tree structure needs to be unique for each word. E.g. "abc"
is not allowed as a word. However, this restriction was already taken into account in generation.
The "choice" variant changed slightly, so that it now contains one wrong answer less and an additional
correct answer. Also, the wrong answers are checked if they are correct by accident.

This change required some refactoring. Some code became unused and was removed.
To increase the readability of the words they are now displayed in markdown as
inline code and the inline code is now displayed with an additional background.